exporting functional component and helper function into seperate files
