+++ frontend/src/modules/Nodes/components/NodeElement/NodeElement.tsx
@@ -12,7 +12,8 @@ import { NodesApi } from "../../api/NodesAPI";
 import { RunIcon } from "../../../../ui-lib/Icons/RunIcon";
 import Tooltip from "@mui/material/Tooltip";
 import { InfoIcon } from "../../../../ui-lib/Icons/InfoIcon";
-import { StatusVisuals, getNodeRowClass } from "./NodeElementStatusVisuals";
+import { StatusVisuals } from "./NodeElementStatusVisuals";
+import { getNodeRowClass } from "./Helpers";
 import { useSnapshotsContext } from "../../../Snapshots/context/SnapshotsContext";
 
 export interface NodeDTO {
