+++ frontend/src/modules/Nodes/components/NodeElement/NodeElementStatusVisuals.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+// eslint-disable-next-line css-modules/no-unused-class
 import styles from "./NodeElement.module.scss";
 import CircularLoaderProgress from "../../../../ui-lib/Icons/CircularLoaderProgress";
 
@@ -11,31 +12,5 @@ export const StatusVisuals: React.FC<{ status?: string; percentage: number }> = 
     return <div className={`${styles.dot} ${styles.redDot}`} />;
   } else {
     return <div className={`${styles.dot} ${styles.greyDot}`} />;
-  }
-};
-
-export const getNodeRowClass = ({
-  nodeName,
-  selectedItemName,
-  lastRunStatusNode,
-}: {
-  nodeName: string;
-  selectedItemName: string;
-  lastRunStatusNode: { name?: string; status?: string } | null;
-}): string => {
-  const isSelected = selectedItemName === nodeName;
-  const isLastRun = lastRunStatusNode?.name === nodeName;
-  const nodeStatus = isLastRun ? lastRunStatusNode?.status : "pending";
-
-  if (isLastRun && nodeStatus === "finished") {
-    return `${styles.rowWrapper} ${styles.nodeSelectedFinished}`;
-  } else if (isLastRun && nodeStatus === "error") {
-    return `${styles.rowWrapper} ${styles.nodeSelectedError}`;
-  } else if (isLastRun && nodeStatus === "running") {
-    return `${styles.rowWrapper} ${styles.nodeSelectedRunning} ${styles.highlightRunningRow}`;
-  } else if (isSelected && nodeStatus === "pending") {
-    return `${styles.rowWrapper} ${styles.nodeSelectedPending}`;
-  } else {
-    return styles.rowWrapper;
   }
 };
